# api.yml
swagger: "2.0"
info:
  description: "APP - TecBikeWeb - Sistema desenvolvido ao longo da disciplinad e TecWeb - UFSJ"
  version: "1.0.0"
  title: "TecBikeWeb"
  contact:
    email: "elder@ufsj.edu.br"
host: "tecbikeweb.ccomp.ufsj.edu.br"
basePath: "/v1"
...
schemes:
- "http"
paths:
  /bike:
    post:
      tags:
      - "bicicleta"
      summary: "Adiciona uma nova bicicleta na locadora"
      operationId: "addBike"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Objeto bicileta a ser adicionada na locadora"
        required: true
        schema:
          $ref: "#/definitions/Bike"
      responses:
        "405":
          description: "Invalid input"
  /bike/{bikeId}:
    get:
      tags:
      - "bicicleta"
      summary: "Busca bicicleta pelo ID"
      operationId: "getBikeById"
      produces:
      - "application/json"
      parameters:
      - name: "bikeId"
        in: "path"
        description: "ID of bike to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Bike"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Bike not found"
    put:
      tags:
      - "bicicleta"
      summary: "UAtualiza uma bicicleta na locadora"
      operationId: "updateBikeById"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "bikeId"
        in: "path"
        description: "ID that need to be updated"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Updated bike object"
        required: true
        schema:
          $ref: "#/definitions/Bike"
      responses:
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Bike not found"
    delete:
      tags:
      - "bicicleta"
      summary: "Apaga uma bicileta"
      description: ""
      operationId: "deleteBike"
      produces:
      - "application/json"
      parameters:
      - name: "bikeId"
        in: "path"
        description: "Bike id to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Bike not found"
...
definitions:
...
  Bike:
    type: "object"
    required:
    - "name"
    properties:
      id:
        type: "integer"
        format: "int64"
      marca:
        type: "string"
      modelo:
        type: "string"
      status:
        type: "string"
        description: "bike status in the store"
        enum:
        - "available"
        - "booked"
        - "rented"